pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deferred-api"
version = "1.0.0"
description = "Deferred — Offline-first payment API. Stripe for the real world."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Deferred Team"}]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis[hiredis]>=5.0.0",
    "aio-pika>=9.4.0",
    "pynacl>=1.5.0",
    "pyjwt[crypto]>=2.8.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "qrcode[pil]>=7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.96.0",
    "locust>=2.20.0",
    "bandit>=1.7.7",
    "ruff>=0.2.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring database",
    "load: marks load tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101", "S104"]

[tool.bandit]
exclude_dirs = ["tests"]
